thingtalk/describe: use primitive templates to describe functions
Replace the use of the `#_[confirmation]` annotation with the
closest match to a primitive template in the dataset (with
a fallback to the canonical form if no template matches).
This allows us to remove the confirmation annotation, which
was redundant and deprecated, and allows us to match what
the state machine would generate.

Also, to make the describe module usable from the state
machine, make it return tokenized sentences, and apply
NLG postprocessing after the fact.