Optimize generation by choosing compatible derivation with a hashtable
Introduce a system of named indices with hashable keys to optimize
the matching of compatible derivations when generating sentences.
The goal is to be able to efficiently sample derivations with
the same context and derivations with compatible function names.
This is necessary to scale up the size of Thingpedia.

Each derivation has a set of keys. Each non-terminal in a template
can be annotated with a key constraint, of the form "name = constant"
or "name1 = other.name2". Only derivations with compatible constraints
are enumerated or sampled, hence we don't need to call the semantic
function with a lot of irrelevant pairs of derivations that cannot
possibly be compatible.
This is implemented efficiently by maintaining a set of hashtable
indices, one for each name declared in the template.

Note that the sampling strategy is unchanged: we keep a fixed
size for each non-terminal, regardless of which indices are used.
But because generation ought to be faster, we should be able to
increase the pruning size (memory limit permitting)

This commit only contains the changes to the generation algorithm.
Actually computing the keys, the syntax to express key constraints,
using contexts as keys, and actually using this functionality in
the ThingTalk templates is left to follow up commits.