Check compatibility of parameters and filters based on function name
Instead of just checking that each function has a parameter of
the given type with the right name, also check that the parameter
comes from an annotation of the right function. That way, we don't
mix and match annotations from different functions for parameters
of the same name, which would lead to a lot of confusion.

The implementation uses a new set of helper objects: ParamSlot,
InputParamSlot, and FilterSlot. These objects augment the ThingTalk
Ast objects by also carrying the FunctionDef of the function that
originated those parameter names/input parameter clauses/filter clauses.
Most of the commit is then adapting the code to those new objects,
and adding isSameFunction checks everywhere.

This implementation slows down synthesis, as it redundant checks of
a lot of incompatible parameters. The next commit will make it fast
by using key constraints.

Fixes #252