Convert placeholder rules to normal rules for result and error phrases
Placeholders are one of the most confusing aspects of the Genie
synthesis algorithm, as they break the intuition of what the grammar
means.
The main reason for their existence was performance, especially when
parameter passing, but that's no longer the case when we have key
constraints. So we can convert them to regular templates when
we load the manifests from Thingpedia.

This commit converts error messages and result phrases, which are
the easy bits because they only deal with constants. The next commit
will deal with the other placeholder-using non-terminals.

Includes some changes to enable testing with test_sentence_generator.