Complete the removal of placeholder-using templates
The last usage of placeholder templates was coreferences. These
are removed by introducing a new internal type, ExpressionWithCoreference.
Such an expression is the result of replacing a placeholder with
"it / this / that" or with "the" + param name. It tracks the
function names and types for efficient matching of the query and action.