Remove the support for placeholders in sentence generation
This completes the clean up, by removing some really ugly code,
and one of the least understandable parts of the Genie template
system. With that gone, templates are now conceptually always
context-free grammar productions
(key constraints are not context-free but those are an optimization only).

"if complete", "[-> pname]", and a whole lot of for loops are gone.