generation: rework how we sample semantic function calls
The sampling strategy of the two generation algorithms are different
in the presence of highly constrained non-terminals. To avoid
calling the semantic function too many times in either algorithm,
we need to keep track of two estimates of productivity, one just
for the semantic function and one for the whole rule.