templates: fix playing songs after an empty spotify search
If a play song command resulted in no songs being played, and
the user retries a different search, we'll preserve the fact
that the user was trying to play a song. But if the user mentions
a name, we should play only the first song matching that name.
This is necessary to be consistent with the initial command, which
otherwise confuses the model quite a bit.

This is a long-standing bug, dating at least to the ThingTalk 2.0
migration, but potentially earlier. It became more visible because
we generate more contexts and more empty searches since the
synthesis refactoring, and that confuses the model even more.