Remove mmap-io for good
I'm seeing segfaults in Travis that are, according to The Google,
related to native modules, and it's either mmap-io or sqlite3.
Likely the former.

The purpose of mmap-io was to shave a few tens of megs of RAM
in almond-cloud. Not a huge deal given that models require gigs
of RAM.