Refactor prediction code for kubeflow (#426)
We don't actually need three separate options for Kubeflow,
all we need is one URL. We can recognize it with a pseudo-scheme
"kf+http(s):" (to distinguish from the RemoteParserClient) and
let the user specify it normally.

Also, refactor Predictor to separate the batching code from
the request code. Remove the ability to have multiple parallel
workers, as multiple workers on the same machine is not very
useful (among other things, they all run on the same GPU), and
move the batching code up. Refactor the Kubeflow backend to be
a straight-forward HTTP call instead of wrapping into a stream.