templates: use describe module to construct confirmations and results
The describe module uses a top-down algorithm to construct a phrase
for a given program, and it's much faster than the bottom-up algorithm
used by normal synthesis, at the expense of not offering any
variety. We'll reintroduce variety later when we refactor the phrase
generation part of the synthesis.

This also lets us confirm arbitrary ThingTalk program, which is a
prereq to support stream, and gets rid of the ugly "your request
completed successfully" in the majority of cases where we can
construct the action phrase and stick the current table in it.