templates: make thingpedia loader not a singleton
Since the migration to TypeScript, singletons declared in
normal .js/.ts files in the templates are process-wide and
persist across multiple rounds of generation. That means
that, at inference time, multiple concurrent generation runs
(which, in almond-cloud, can belong to different users!)
would share the same thingpedia loader, stomp on each other,
and make a mess.

This is obviously not acceptable, so we need to initialize
a new separate Thingpedia loader for every generation run.
This means we sadly need to reintroduce the magic $loader
variable we used to have a long time ago, and we need to
pass it around everywhere.

Spotted by chance while working on something else. Given
it's a concurrency bug, had I not noticed it, we would have
cried debugging this in production.