Implement FAQ based meta-commands (#430)
Meta commands such as "What is your name" or "what can you do"
are implemented using an action in the thingengine.builtin interface
(aka the "Miscellaneous" skill), which is always available. The
action takes an enum mapping the intent of the question, and returns
the actual reply.

I chose an action instead of a query because we reset the state
of the dialogue after a successful action, so there is less risk
of carrying over weird state between meta questions. Additionally,
actions in the builtin interface are blocked from normal templates,
and only use the "program" primitive templates. That way we don't generate
a weird "get reply to ${question}" where question is an enum, which
would pollute the training set.

To ease development of the meta-skill, the actual content is
specified in a faq.yaml file, and the regular Thingpedia manifest
is auto-generated.