Restore detailed message when no device of a given type is configured
The generic message is confusing, because it sounds like the agent
doesn't know what the command is about, and the word "skill" is
confusing. We used to need the generic message because we could
not distinguish a bad parse / OOD command from a good parse without
the right skill. We have confidence for that now, so we can be
confident that the command was parsed correctly.

Fixes #450