Transform the program to add [1] clause before execution
We need to add the [1] clause to implement the right semantics
for "play X by Y" on Spotify, as there can be many songs with
the same or similar name. We want ThingTalk involved to
choose the right one, but we only want to play one song. So
we use the same heuristic we had before during synthesis, but
we apply it before execution (including simulation, so there
is no training/test mismatch)