Disable support for multidomain contexts
This is a large hammer to solve some of the batch size issues we're
seeing in the Thingpedia model, as we generate contexts that are
too long. With this change, the context will contain at most
one statement, and therefore will be quite short, at the expense
of not supporting any multi-domain parameter passing.

This commit is incompatible with MultiWOZ and will need to be
reverted or made into an option.