builtin: improve logic to choose the default bounds for random (#475)
If only one bound is specified, use a more clever logic to choose
the other bound, accounting for the sign of the bound. In the positive
case, if low is specified, high is set to 2*low; if high is specified,
low is set to 1. The negative case is symmetric.

Fixes #474