Make use of the new templates for the agent replies
This is an initial pass converting Genie templates to the new
syntax. It is enough to cover most of the agent replies, and it
solves all the grammar mistakes in the agent tests.

The old hack to "fix" the grammar in post-processing is gone,
as it was doing more harm than good.