Clean up card objects (rich messages)
Remove unused/unimplemented ones, split audio and video messages,
and merge the implementation of audio, video, and pictures.
Also, remove the catch-all result message. It adds a layer of
indirection that is not necessary. If a platform does not implement
a certain type of card it can just ignore it, because the main
result phrase should contain enough information.