Rework how sound effects are played
Add two modes: exclusive, and non exclusive. In non-exclusive mode,
we play the sound effect at the end of the speech, without synchronizing
with background audio. In exclusive mode, we treat the sound effect
like an audio message, so it stops other audio and synchronizes
with audio messages from the same frame.