Add support for lazy (async iterable) queries
This is the last bit of infrastructure to support lazy queries,
that is, queries implemented using `async *get_foo` which yield
their results lazily instead of return an array with all the results.

The exec environment caches the output of queries for a given
statement. The cache must be an iterable that can be iterated multiple
times, rather than a generator, which can be iterated only once.
To this end, we introduce a new class that wraps an iterable
and queries it lazily, but also caches each returned item so
iteration can be restarted.

Fixes #238