Refactor sqlite interfaces, abstract them out to support other platforms
Move all sqlite specific code to lib/engine/db/sqlite, and introduce
abstract interfaces in lib/engine/db that provides a type-safe,
DB-agnostic way for other components in the engine to access the
database. This will allow to plug other DB backends in the future,
such as an external DB shared by multiple users in the cloud.