Load types from bootleg and use it as a reference to decide value types
Each value has to be assigned to a single entity type, otherwise, we
will have program containing the same entity but with different type
info: since we now allow multiple inheritance, there might be multiple
types typecheck for the same property.

To detemine the "best" type for each entity, we try to be consistent
with bootleg output, where we normally pick the 1st one or two types
extracted in bootleg.