Move templates into lib/templates
With this change, templates become part of Genie proper, instead
of being kind of a separate package that uses Genie. This is
to reflect the fact that we don't expect third parties to write
templates in the future, because, especially with dialogues, templates
are just too coupled with the core Genie and it doesn't make sense
to develop them separately.
Removing the distinction allows us to clean the build system
and remove the confusion of having two build directories.