Refactor how confidence scores are computed
Use exactly one score to determine if a parse is likely correct,
and exactly one score to determine if a parse is likely in-domain.

This corresponds exactly to the dispatch interface, it is closer
to the low-level confidence estimators, and it's overall cleaner.

In the API, we still expose a multi-class classifier for the
"intent" of the command (which we can sub-out in the future),
and a score on each parse.