Wire up the new dialogue generator for synthesis
Extract the dialogue generation code into a new module, and
wire up the interfaces used by DialogueInterface in a new
PartialDialogue class.

The IO of DialogueInterface is rewritten in terms of templates,
so we can integrate with the existing synthesis. These templates
are added dynamically to the user/agent generator, which are
then expanded normally.