Refactor the synthesis algorithm
In preparation for having dialogue PL call the synthesis algorithm
with arbitrary templates at any time, we need to make the synthesis
algorithm dynamic: instead of bottom-up construct every possible
derivation towards the root, the derivations are constructed
top-down based of what is needed by the root non-terminal.

The same chart system is used for pruning, so this effectively
converts the algorithm from a dynamic programming to a recursion
with memoization, which should have the same computational complexity
and generate the same output.

There are two functional changes in this commit:
- the order in which rules are applied is different: before, we would
  proceed depth-wise, now the order in unpredictable because it
  depends on how we explore the derivation tree
- sampling for the root non-terminal is different: before there was no
  sampling for the root non-terminal, now the root non-terminal is treated
  like any other non-terminals and subject to sampling