Introduce the notion of policy modules
Consolidate the new policy functions, the policy manifest, and
the initialization of the templates into a single JS file.
This takes advantage of the fact that compiled templates are just plain
TS files that can be imported normally.

The "--template-file" flag to "generate-dialogs" is removed and
replaced with an optional "--policy-module" that specifies the
path to the policy module JS file. If that's unspecified, the
transaction policy is used.

No changes to single-sentence generation, which can continue to
use custom template files if desired.