templates: handle notifications for timers without functions correctly
If the user issues a command that gets parsed into a timer without
another Thingpedia function, the schema of the command will just
be the timer. This command does not have a class, and it is not
an Ast.Invocation.

Handle that case gracefully everywhere, even if it is an erroneous
program.

Fixes #695