Fix simulating relative locations
If the simulated location is called "here", when the sentence contains
the word "here" for a different filter, the location will be extracted
from the sentence and not from the context when converting to contextual
format. This is incorrect and results in two sentences with the
same context and two different parses.