Ban follow-up commands by name
In multiwoz, a follow-up command in the same domain that uses a
name refers to that name in context, and carries over all the slots.

In production, a command with a name is a full command that should
discard the previous context. Only commands with pronoun coreferences
should carry over the slots.

This is causing inconsistency in training set since the recent
changes, and weird bugs due to resolving a name from a context that
doesn't have that name.