Pass thingtalk version to Thingpedia API calls in download-* commands
The Thingpedia API uses the ThingTalk version to activate API
compatibility, where needed. If we don't pass it, it will assume
the latest version of ThingTalk and break.

Most recently, this broke because of multiple entity inheritance,
as Thingpedia started returning an array for "subtype_of" instead
of a single string or null.