utils/thingtalk: handle dialogue annotations correctly when cloning states
Remove all annotations when computing a prediction (delta between two states)
and remove all annotations when preparing a context for input to the neural
model. This means annotations are only preserved in the C: state, and only
in JS land (never in the neural model), which is the intended behavior.