Always use Bootleg types when available
Each entity will be assigned to the _fisrt_ type in bootleg, so it's
always consistent with the bootleg prediction at inference type.
This is possible because the type is not useful when converting
ThingTalk to SPARQL, it's only used to assist parsing. Thus, we don't
need to be consistent with CSQA.

In case when Bootleg does not have the information of the type
information, we fallback to the CSQA type - which is less than 1% in
practice (based on result in city domain).