Remove trailing QIDs in bootleg type canonical
When there are multiple types share the same canonical, bootleg will
append QID at the end of the canonical for all types except one.
Sometimes, it also append QID to some type that does not share
canonical with other types, such as "nation", and "designation", not
sure why.

In our case, the type information is to assist the parsing, the actual
QID is not important, so we drop all the appended QIDs - if two types
share the same canonical, they are considered as the same type in
natural language.