Add conversation protocol to synchronize devices
In genie-client-cpp, we need to retrieve the Spotify credentials
to pass to spotifyd. We could do that over the cloud sync channel,
but to avoid opening another connection, let's add a new (optional)
message type to the conversation that reports any newly configured
device with the full set of credentials.