Add templates for direct result phrases with array parameters (#748)
"Direct result phrases" are phrases of the form "The X is Y."

Add the variants of these phrases for arrays, so we can query
the artists of the currently playing song. We need this template
because we don't use the form "I found X. It is a song written by Y" if
the query is not a list.