Use timezone-aware code when comparing and displaying dates
Always operate in the timezone of the user.

This gets one step closer to a successful run of test_describe
with a system timezone (TZ environment) that does not match
the user timezone (always America/Los_Angeles in tests).
The missing bits are due to ThingTalk parsing date/time syntax
incorrectly.