audio controller: add the ability to control multiple players
Introduce new interfaces for abstract "players" which can be
local direct playback (using platform capabilities aka GStreamer
+PulseAudio) or remote playback controlled over the conversation
web socket.

Track one abstract player for each conversation, and the skill
associated with that player. Use that information to know which
player should be stopped when a command comes.
Skills are expected to thread through the conversation ID from
the exec environment to the API calls they make to the audio
controller.