templates: distinguish between monitoring the whole row and an item in the row
If the user asks for "when there are new X", make that a monitor
on the ID (if X is a list and has an `id` argument), so unrelated
changes to the current rows of X won't cause a notification.

Conversely, when the user asks for "when the X change", we monitor
and notify on everything.