Allow overriding the canonical forms of enum values
Using the "enum_value" or "value_enum" keys of the #_[canonical]
annotation of a parameter that uses the type.

Because enums are used interchangeably on all parameters with the
same type, the same canonical must be used on all parameters,
and it is unspecified which one will be used otherwise.